faster way to get student absence totals
