+++ app/assets/javascripts/school_administrator_dashboard/main.jsx
@@ -1,9 +1,9 @@
-import SchoolwideAttendance from  './schoolwide_attendance.jsx';
+import SchoolwideAbsences from  './schoolwide_absences.jsx';
 import MixpanelUtils from '../helpers/mixpanel_utils.jsx';
 
 export default function renderSchoolAdminDashboardMain(el) {
   const serializedData = $('#serialized-data').data();
   MixpanelUtils.registerUser(serializedData.currentEducator);
   MixpanelUtils.track('PAGE_VISIT', { page_key: 'SCHOOL_DASHBOARD' });
-  window.ReactDOM.render(<SchoolwideAttendance dashboardStudents={serializedData.students} />, el);
+  window.ReactDOM.render(<SchoolwideAbsences dashboardStudents={serializedData.students} />, el);
 }
