+++ app/assets/javascripts/school_administrator_dashboard/schoolwide_attendance.jsx
@@ -1,28 +1,62 @@
 import React from 'react';
 
 import DashboardHelpers from './dashboard_helpers.jsx';
-import PageContainer from './page_container.jsx';
+import SchoolAbsenceDashboard from './school_absence_dashboard.jsx';
 
 export default React.createClass({
   displayName: 'SchoolwideAttendance',
 
   propTypes: {
-    dashboardStudents: React.PropTypes.array.isRequired  //should rename more generically
+    schoolAbsenceEvents: React.PropTypes.object.isRequired,
+    dashboardStudents: React.PropTypes.array.isRequired
   },
 
   schoolAverageDailyAttendance: function() {
-    const studentRecords = this.props.dashboardStudents;
-    return DashboardHelpers.averageDailyAttendance(DashboardHelpers.absenceEventsByDay(studentRecords), this.totalStudents());
+    return DashboardHelpers.averageDailyAttendance(this.props.schoolAbsenceEvents, this.props.dashboardStudents.length);
   },
 
-  totalStudents: function() {
-    return this.props.dashboardStudents.length;
+  homeroomAverageDailyAttendance: function() {
+    const studentRecords = this.props.dashboardStudents;
+    const studentsByHomeroom = DashboardHelpers.groupByHomeroom(studentRecords);
+    const eventsByHomeroom = this.homeroomAbsenceEventsByDay(studentsByHomeroom);
+    let homeroomAverageDailyAttendance = {};
+    Object.keys(eventsByHomeroom).forEach((homeroom) => {
+      const homeroomSize = studentsByHomeroom[homeroom].length;
+      homeroomAverageDailyAttendance[homeroom] = DashboardHelpers.averageDailyAttendance(eventsByHomeroom[homeroom], homeroomSize);
+    });
+    return homeroomAverageDailyAttendance;
+  },
+
+  homeroomAbsenceEventsByDay: function(studentsGroupedByHomeroom) {
+    let homeroomAbsenceEventsByDay = {};
+    Object.keys(studentsGroupedByHomeroom).forEach((homeroom) => {
+      const daysWithAbsences = DashboardHelpers.absenceEventsByDay(studentsGroupedByHomeroom[homeroom]);
+      homeroomAbsenceEventsByDay[homeroom] = this.addPerfectAttendanceDays(daysWithAbsences);
+    });
+    return homeroomAbsenceEventsByDay;
+  },
+
+  //Because homerooms often have no absences, merge their daily events with the list of school days
+  addPerfectAttendanceDays: function(eventsGroupedByDay) {
+    this.schoolDays().forEach((day) => {
+      if (!eventsGroupedByDay[day]) eventsGroupedByDay[day] = [];
+    });
+    return eventsGroupedByDay;
+  },
+
+  //There's no application awareness of valid school days, but there are almost never schoolwide perfect attendance days
+  //We use days the school has recorded at least one absence as a proxy for school days
+  schoolDays: function () {
+    return Object.keys(this.props.schoolAbsenceEvents);
   },
 
   render: function() {
     return (
-        <PageContainer
+        <SchoolAbsenceDashboard
           schoolAverageDailyAttendance = {this.schoolAverageDailyAttendance()}
-          dashboardStudents = {this.props.dashboardStudents}/>);
+          homeroomAverageDailyAttendance = {this.homeroomAverageDailyAttendance()}
+          schoolAbsenceEvents = {this.props.schoolAbsenceEvents}
+          dashboardStudents = {this.props.dashboardStudents}
+          dateRange = {Object.keys(this.schoolAverageDailyAttendance()).sort()}/>);
   }
 });
